pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pisama-auto"
version = "0.1.0"
description = "Zero-code auto-instrumentation for LLM applications — adds Pisama failure detection with one line"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pisama Team", email = "team@pisama.dev" }
]
keywords = [
    "ai", "agents", "monitoring", "auto-instrumentation",
    "opentelemetry", "observability", "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "wrapt>=1.15.0",
]

[project.urls]
Homepage = "https://github.com/tn-pisama/pisama-auto"
Documentation = "https://docs.pisama.com/auto-instrumentation"

[tool.hatch.build.targets.wheel]
packages = ["src/pisama_auto"]
